pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
description = ""
[tool.poetry.dependencies]
python = "^3.8"
Flask = "^3.1.2"
Flask-Session = "^0.8.0"
Flask-SQLAlchemy = "^3.1.1"
Flask-Migrate = "^4.0.7"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.0.1"
requests = "^2.32.5"
